move `FakesRoles` into `Statamic\Testing\Concerns`
so addons using `AddonTestCase` can build users with and without a
given permission, instead of hand-copying the trait.
